requirements.txt
albumentations==1.2.0
deta==1.1.0
grad_cam==1.3.9
matplotlib==3.5.2
numpy==1.22.4
opencv-contrib-python-headless
pandas==1.4.2
Pillow==9.1.1
python-dotenv==0.20.0
requests==2.28.0
seaborn==0.11.2
streamlit==1.10.0
streamlit_authenticator==0.1.5
streamlit_cropper==0.2.0
streamlit_lottie==0.0.3
timm==0.5.4
torch==1.11.0
torchvision==0.12.0
